# Improve formatter performance