… Makefile
… go.sum