Use [valyala/bytebufferpool](https://github.com/valyala/bytebufferpool) to reduce memory wast during formatting log

```
goos: darwin
goarch: amd64
pkg: github.com/gol4ng/logger/formatter
cpu: Intel(R) Core(TM) i5-7267U CPU @ 3.10GHz
BenchmarkFormatter_Format
BenchmarkFormatter_Format/default_test_message
BenchmarkFormatter_Format/default_test_message-4         	 8734137	       137.4 ns/op	      56 B/op	       3 allocs/op
BenchmarkFormatter_Format/default_colored_test_message
BenchmarkFormatter_Format/default_colored_test_message-4 	 7569991	       156.1 ns/op	      56 B/op	       3 allocs/op
BenchmarkFormatter_Format/default_context_test_message
BenchmarkFormatter_Format/default_context_test_message-4 	  936051	      1438 ns/op	     584 B/op	      11 allocs/op
BenchmarkFormatter_Format/default_full_test_message
BenchmarkFormatter_Format/default_full_test_message-4    	  863792	      1380 ns/op	     584 B/op	      11 allocs/op
BenchmarkFormatter_Format/gelf_test_message
BenchmarkFormatter_Format/gelf_test_message-4            	  507562	      2326 ns/op	    1376 B/op	      18 allocs/op
BenchmarkFormatter_Format/json_test_message
BenchmarkFormatter_Format/json_test_message-4            	  817722	      1684 ns/op	     904 B/op	      13 allocs/op
BenchmarkFormatter_Format/line_test_message
BenchmarkFormatter_Format/line_test_message-4            	 1159873	       943.4 ns/op	     392 B/op	       7 allocs/op
```

```
goos: darwin
goarch: amd64
pkg: github.com/gol4ng/logger/formatter
cpu: Intel(R) Core(TM) i5-7267U CPU @ 3.10GHz
BenchmarkFormatter_Format
BenchmarkFormatter_Format/default_test_message
BenchmarkFormatter_Format/default_test_message-4         	11890204	       100.9 ns/op	      24 B/op	       1 allocs/op
BenchmarkFormatter_Format/default_colored_test_message
BenchmarkFormatter_Format/default_colored_test_message-4 	10601418	       113.2 ns/op	      32 B/op	       1 allocs/op
BenchmarkFormatter_Format/default_context_test_message
BenchmarkFormatter_Format/default_context_test_message-4 	 1099491	      1089 ns/op	     224 B/op	       6 allocs/op
BenchmarkFormatter_Format/default_full_test_message
BenchmarkFormatter_Format/default_full_test_message-4    	 1000000	      1098 ns/op	     224 B/op	       6 allocs/op
BenchmarkFormatter_Format/gelf_test_message
BenchmarkFormatter_Format/gelf_test_message-4            	  567135	      1884 ns/op	     548 B/op	       8 allocs/op
BenchmarkFormatter_Format/json_test_message
BenchmarkFormatter_Format/json_test_message-4            	 1000000	      1069 ns/op	     240 B/op	       6 allocs/op
BenchmarkFormatter_Format/line_test_message
BenchmarkFormatter_Format/line_test_message-4            	 1571288	       770.6 ns/op	     256 B/op	       3 allocs/op
```